Fix ppc64le backend and 64K page size support (#2828)
- Fix `ppc64le` backend (`ppc64` and 32-bit 
  PowerPC still broken)
- Autodetect 4K or 64K pages on `ppc64` and 
  `ppc64le`
- Make `PAGESIZE` a `uint32_t` if defined 
  (because 64K won't fit in `uint16_t`)
- Since `mmap` is page aligned by definition, 
  don't try to realign allocations we get from it 
  (the current alignment code strips the top 
  word off 64-bit pointers)
- Alternate code for `dyn_mem_adjust` to properly handle 64K pages

---------

+++ include/support.h
@@ -52,7 +52,8 @@
 #endif
 
 #ifdef PAGESIZE
-constexpr uint16_t host_pagesize = { PAGESIZE };
+// Some platforms like ppc64 have page sizes of 64K, so uint16_t isn't enough.
+constexpr uint32_t host_pagesize = { PAGESIZE };
 #else
 constexpr uint16_t host_pagesize = 4096;
 #endif
